# slices: inline search into its callers