--- a/slices/sort_benchmark_test.go
+++ b/slices/sort_benchmark_test.go
@@ -5,6 +5,7 @@
 package slices
 
 import (
+	"fmt"
 	"math/rand"
 	"sort"
 	"strings"
@@ -198,5 +199,22 @@ func BenchmarkSortFuncStructs(b *testing.B) {
 		ss := makeRandomStructs(N)
 		b.StartTimer()
 		SortFunc(ss, lessFunc)
+	}
+}
+
+func BenchmarkSearchFloats(b *testing.B) {
+	for _, size := range []int{16, 32, 64, 128, 512, 1024} {
+		b.Run(fmt.Sprintf("Size%d", size), func(b *testing.B) {
+			floats := make([]float64, size)
+			for i := range floats {
+				floats[i] = float64(i)
+			}
+			midpoint := len(floats) / 2
+			needle := (floats[midpoint] + floats[midpoint+1]) / 2
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				BinarySearch(floats, needle)
+			}
+		})
 	}
 }