--- a/slices/sort_benchmark_test.go
+++ b/slices/sort_benchmark_test.go
@@ -202,7 +202,7 @@ func BenchmarkSortFuncStructs(b *testing.B) {
 	}
 }
 
-func BenchmarkSearchFloats(b *testing.B) {
+func BenchmarkBinarySearchFloats(b *testing.B) {
 	for _, size := range []int{16, 32, 64, 128, 512, 1024} {
 		b.Run(fmt.Sprintf("Size%d", size), func(b *testing.B) {
 			floats := make([]float64, size)
@@ -214,6 +214,24 @@ func BenchmarkSearchFloats(b *testing.B) {
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
 				BinarySearch(floats, needle)
+			}
+		})
+	}
+}
+
+func BenchmarkBinarySearchFuncStruct(b *testing.B) {
+	for _, size := range []int{16, 32, 64, 128, 512, 1024} {
+		b.Run(fmt.Sprintf("Size%d", size), func(b *testing.B) {
+			structs := make([]*myStruct, size)
+			for i := range structs {
+				structs[i] = &myStruct{n: i}
+			}
+			midpoint := len(structs) / 2
+			needle := &myStruct{n: (structs[midpoint].n + structs[midpoint+1].n) / 2}
+			lessFunc := func(a, b *myStruct) int { return a.n - b.n }
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				BinarySearchFunc(structs, needle, lessFunc)
 			}
 		})
 	}