Constructing a lambda and calling the search function has shown to not
perform well, being on par with the non-generic sort.Search. Inlining
gives a major performance boost

```
~/exp$ benchstat gotip_master.bench gotip_inlined.bench
name                     old time/op  new time/op  delta
SearchFloats/Size16-4    20.1ns ± 6%   6.2ns ± 1%  -69.24%  (p=0.000
n=9+9)
SearchFloats/Size32-4    22.2ns ± 3%   7.2ns ± 1%  -67.69%  (p=0.000
n=9+9)
SearchFloats/Size64-4    25.8ns ± 6%   8.2ns ± 1%  -68.36%  (p=0.000
n=9+10)
SearchFloats/Size128-4   28.9ns ± 7%   9.2ns ± 0%  -68.30%  (p=0.000
n=10+10)
SearchFloats/Size512-4   34.8ns ± 6%  11.2ns ± 0%  -67.83%  (p=0.000
n=10+10)
SearchFloats/Size1024-4  36.9ns ± 3%  12.2ns ± 0%  -66.99%  (p=0.000
n=9+9)
```

The story is similar for BinarySearchFunc, even though less severe

```
~/exp$ benchstat gotip_master_structs.bench gotip_inlined_structs.bench 
name                            old time/op  new time/op  delta
BinarySearchFuncStruct/Size16-4    28.9ns ± 2%  19.0ns ± 4%  -34.18%  (p=0.000 n=9+10)
BinarySearchFuncStruct/Size32-4    33.8ns ± 1%  22.0ns ± 5%  -35.03%  (p=0.000 n=9+10)
BinarySearchFuncStruct/Size64-4    39.0ns ± 1%  23.4ns ± 1%  -39.98%  (p=0.000 n=10+9)
BinarySearchFuncStruct/Size128-4   43.8ns ± 1%  26.0ns ± 0%  -40.78%  (p=0.000 n=9+8)
BinarySearchFuncStruct/Size512-4   54.0ns ± 0%  31.5ns ± 1%  -41.58%  (p=0.000 n=10+10)
BinarySearchFuncStruct/Size1024-4  59.0ns ± 0%  34.1ns ± 0%  -42.30%  (p=0.000 n=9+9)
```